requirements.txt
torch>=1.10.0
torchvision>=0.11.0
timm>=0.6.0
h5py>=3.6.0
pandas>=1.3.0
Pillow>=9.0.0
numpy>=1.19.0